Change default enum to be INTEGER, use that in code examples and tests

+++ AerospikeDemo/QueryResume.cs
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2023 Aerospike, Inc.
+ * Copyright 2012-2026 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -90,7 +90,7 @@ namespace Aerospike.Demo
 
 			try
 			{
-				IndexTask task = client.CreateIndex(policy, args.ns, setName, indexName, binName, IndexType.NUMERIC);
+				IndexTask task = client.CreateIndex(policy, args.ns, setName, indexName, binName, IndexType.INTEGER);
 				task.Wait();
 			}
 			catch (AerospikeException ae)
